Projet_5DLA_Calixte_NGUEMO_et_Corelia_Wildervia/requirements.txt
setuptools>=65.0
mlflow==1.30.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
imbalanced-learn==0.10.1
matplotlib==3.6.3
seaborn==0.12.2
pyyaml==6.0
xgboost==1.7.6